16. In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A+

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,21 +6,21 @@
  */
 int main(void)
 {
-	unsigned long long int a = 1, b = 2, c,  count = 0;
+	unsigned long int a = 1, b = 2, c,  count = 0;
 
-	printf("%lld, ", a);
+	printf("%ld, ", a);
 	count++;
-	printf("%lld, ", b);
+	printf("%ld, ", b);
 	count++;
 	while (count < 98 - 1)
 	{
 		c = a + b;
 		a = b;
-		printf("%lld, ", c);
+		printf("%ld, ", c);
 		b = c;
 		count++;
 	}
-	printf("%lld", c);
+	printf("%ld", c);
 	printf("\n");
 	return (0);
 }
